extract org user & event routes into subrouters
